Task Error

+++ main.py
@@ -1,5 +1,8 @@
 from flask import *
-from flask_sqlalchemy import SQLAlchemy
+from flask_sqlalchemy import SQLAlchemy, sqlalchemy
+from sqlalchemy.exc import SQLAlchemyError
+
+# from sqlalchemy import SQLAlchemy
 
 app = Flask(__name__)
 app.config["JSON_AS_ASCII"] = False
@@ -81,11 +84,11 @@ test_json3 = {
     "internship_id": 1,
     "memo": "IT系、結構よさげ",
     "priority": 1,
+    "tasks": [
+        {"task_id": 3, "task_name": "応募", "start_date": "2021/08/27 00:00", "end_date": "2021/08/27 00:00", "is_complete": False},
+        {"task_id": 4, "task_name": "面接本番", "start_date": "2021/08/27 00:00", "end_date": "2021/08/28 00:00", "is_complete": True}
+    ]
 }
-    # "tasks": [
-    #     {"task_id": 0, "task_name": "応募", "start_date": "2021/08/27 00:00", "end_date": "2021/08/27 00:00", "is_complete": False},
-    #     {"task_id": 1, "task_name": "面接本番", "start_date": "2021/08/27 00:00", "end_date": "2021/08/28 00:00", "is_complete": True}
-    # ]
 
 test_json4 = {
     "req_num": 4,
@@ -116,6 +119,7 @@ def request_from_client():
         # data = [ json.loads(s) for s in data if s != "" ]
         # json_from_client = json.loads(str(data))
         json_from_client = test_json
+        app.logger.info('json_from_client: %s', json_from_client)
         req_num = json_from_client["req_num"]
         if req_num == 0:
             json_to_client = create_user(json_from_client)
@@ -178,23 +182,46 @@ def update_internship(json):
     internship_id = json["internship_id"]
     memo = json["memo"]
     priority =json["priority"]
-    # tasks = json["tasks"]
+    tasks = json["tasks"]
     
     #新規作成
     if internship_id == -1:
        internship_id = create_internship_id(user_id, internship_id)
+    app.logger.info('internship_id: %s', internship_id)
     ###ここまで成功
 
     #制限して取得
-    my_internship = db.session.query(Internship).filter(Internship.user_id==user_id, Internship.id==internship_id)
+    my_internship = db.session.query(Internship).filter(Internship.user_id==user_id, Internship.id==internship_id).one()    
+    app.logger.info('Old')
+    app.logger.info('my_internship: %s', my_internship)
+
+    # app.logger.info('my_internship.memo: %s', my_internship.memo)
+    # app.logger.info('my_internship.priority: %s', my_internship.priority)
 
     #DBの書き換え
-    my_internship.memo = "メモテスト"
-    my_internship.priority = 3
-    # my_internship.tasks = tasks
+    my_internship.memo = memo
+    my_internship.priority = priority
     # task_list = update_tasks(user_id, internship_id, tasks)
+    
+    try:
+        task_list = update_tasks(user_id, internship_id, tasks)
+    except sqlalchemy.orm.exc.UnmappedInstanceError as e:
+        # error = str(e.__dict__['orig'])
+        # app.logger.info('error: %s', error)
+        app.logger.info('e: %s', e)      
 
+    app.logger.info('my_internship.memo: %s', my_internship.memo)
+    app.logger.info('my_internship.priority: %s', my_internship.priority)
+
+    app.logger.info('New')
+    app.logger.info('my_internship: %s', my_internship)
+    # try:
     db.session.add(my_internship)
+    # except SQLAlchemyError as e:
+    #     # error = str(e.__dict__['orig'])
+    #     # app.logger.info('error: %s', error)
+    #     app.logger.info('e: %s', e)       
+        
     db.session.commit()
     db.session.close()
 
@@ -205,43 +232,69 @@ def update_internship(json):
     return new_json
 
 def create_internship_id(user_id, internship_id):
+
     new_internship = Internship(user_id=user_id)
     db.session.add(new_internship)
+    
     db.session.commit()
     db.session.close()
     internship_id = db.session.query(Internship).filter(Internship.user_id==user_id).first().id
     return internship_id
 
-""""
 def update_tasks(user_id, internship_id, tasks):
-    tasks_db = db.session.query(Task).filter(Internship.user_id==user_id, Internship.id==internship_id)
-    db.session.delete(tasks_db) 
-    db.session.commit()
+    app.logger.info('user_id: %s', user_id)
+    app.logger.info('internship_id: %s', internship_id)
+    app.logger.info('tasks: %s', tasks)
+
+    tasks_db = db.session.query(Task).filter(Task.internship_id==internship_id).all()
+    if tasks_db is None:
+        app.logger.info('tasks_db_None: %s', tasks_db)
+    else:
+        app.logger.info('tasks_db_Exist: %s', tasks_db)
+        #TODO
+        try:
+            db.session.delete(tasks_db) 
+            # ここエラーでる
+        except :
+            db.session.commit()
+        #ここまでOK
 
     json_task_id_list = [task["task_id"] for task in tasks]
+    app.logger.info('json_task_id_list: %s', json_task_id_list)
+    # ここまでOK
+
     # for i in range(json_task_id_list[-1] + 1):
-    for i in json_task_id_list:
-      new_task = Task(internship_id=internship_id)
-      db.session.add(new_task)
-      db.session.commit()
+    for num, id in enumerate(json_task_id_list):
+        new_task = Task(internship_id=internship_id,id=id) #この書き方は大丈夫？
+    #   new_task = Task()
+        app.logger.info('new_task: %s', new_task)
+        app.logger.info('ここから心配')
+    #   db.session.add(new_task)
+    #   db.session.commit()
 
-      new_task = tasks_db.first()
+    #   new_task = db.session.query(Task).filter(Task.internship_id==internship_id).first() #書き換えとして扱う
+    # idが書き換え不可能だからこうしてる？ 分岐させるのみ？
 
-      new_task.id = [task["task_id"] for task in tasks]
-      new_task.name = [task["task_name"] for task in tasks]
-      new_task.start_date = [task["start_date"] for task in tasks]
-      new_task.end_date = [task["end_date"] for task in tasks]
-      new_task.is_complete = [task["is_complete"] for task in tasks]
-      db.session.commit()
-      db.session.close()
-    return [task.to_dict for task in Task.query().filter(Internship.user_id==user_id, Internship.id==internship_id)]
+    #   new_task.id = [task["task_id"] for task in tasks] # id書き換え可能なのか？
+    #   new_task.internship_id = [task["internship_id"] for task in tasks]
+        new_task.name = tasks[num]["task_name"]
+        new_task.start_date = tasks[num]["start_date"]
+        new_task.end_date = tasks[num]["end_date"] 
+        new_task.is_complete = tasks[num]["is_complete"] 
+        app.logger.info('new_task.id : %s', new_task.id )
+        app.logger.info('new_task.name: %s', new_task.name)
+        app.logger.info('new_task.start_date: %s', new_task.start_date)
+        app.logger.info('new_task.end_date: %s', new_task.end_date)
+        app.logger.info('new_task.is_complete: %s', new_task.is_complete)
 
-    # db_task_id_list = [task.id for task in tasks_db]
-    # edit_list = []
-    # create_list = []
-    # delete_list = []
-    #   for j in db_task_id_list:
-"""
+        db.session.add(new_task)
+        app.logger.info('new_task: %s', new_task)
+        #ここからか！
+        try:
+            db.session.commit()
+        except:
+            db.session.close()
+    return [task.to_dict for task in Task.query().filter(Task.internship_id==internship_id).all()]
 
 def delete_internship(json):
     user_id = json["user_id"]
